Site Editor Keyboard Navigation Flow (#50296)
* Add focus ring and keydown event to enter site editor

The site editor is an iframe, which doesn't reliably emit focus events. To work around this, when the blur and focus events are emitted on the `useTabNav` from `useWritingFlow` on the iframe editor, the iframe component can set a `.is-focused` classname on the iframe to show a visual outline. The onClick event listener doesn't work on the iframe to handle a spacebar or enter keypress, so we had to add an onKeyDown listener for this as well.

* Set W logo as a Go back to dashboard link when site editor navigation is open

We need the W logo element tag to stay the same between both the navigation open and closed states so that the component doesn't remount and beak the animation. To do this, we have to set an `href` to force the <Button> component to stay as an <a>.

* Use CSS var for border width focus and calc

* Fix missing - (negative) in the offset

* refactor to remove many conditionals

* Refactor how to pass props to Iframe

* Lint js

---------

Co-authored-by: scruffian <ben@scruffian.com>