Make getBlockEditingMode() return 'default' when parent is 'contentOnly' (#51185)
Change the behaviour of getBlockEditingMode() so that a block's children
is considered "content" that is editable when the editing mode is
'contentOnly'. Therefore getBlockEditingMode( foo ) should be 'default'
when getBlockEditingMode( getParent( foo ) ) is 'contentOnly'.

The benefit of this is that block subtrees can be enabled by simply
setting the parent container to 'contentOnly' rather than each
individual block to 'default'. This is much more performant as each call
to setBlockEditingMode() invalidates the cache.