Image block: Add animation toggle to lightbox behavior (#51357)
* Remove reset button in behaviors, use dropdown option instead

* Use default option reading from the theme, prevent autoupdate

* Add zoom animation UI and styles

* Add logic to use original image and scale its dimensions

Because the image in the lightbox has absolute positioning and
does not respect the padding of its parent container, we need to
get references to both the parent <figure> element and the <image>
element to set the right scale and smoothly animate the zoom.

To accomplish that, since we don't have access to the img src
or its natural width and height until it actually appears
in the DOM, I needed to hoist the imgSrc up to the parent context
to allow for retrieval of the target dimensions from an <img>
element created on the fly.

* Remove extraneous help text

* Manually center lightbox image to improve animation performance

The previous method of centering the image was peforming poorly
on mobile. By doing more manual calculation, the animation now
performs much better.

* Move and reenable class declaration for overflow

The 'has-lightbox-open' class was previously causing the
content to shift before the image animation occurred and looked
like a mistake. I've now moved the declaration so that the
class is added during the animation so it draws less attention.

* Add prefers reduced motion accessibility styles

* Modify fade styles to prevent image flashing

* Simplify code for lightbox UI

* Fix PHP error and linter syntax

* Clean up code; fix bug, add comments

Mostly moved code around, renamed variables for clarity, and
add comments.

Fixed a bug wherein the lightbox wouldn't close on scroll
when using a fade animation.

* Fix bug wherein newly placed images were not setting lightbox animation

* Fix bug wherein vertical images were stretched on mobile

Removed stylesheet padding declarations for the lightbox
and cleaned up logic to ensure correct dimensions get set
for vertical images on mobile devices.

* Update e2e tests

* Update e2e tests, fix selector showing when it should not

---------

Co-authored-by: Carlos Bravo <carlos.bravo@automattic.com>