Perf: Rely on a single store listener only
Currently when a component calls useSelect or any of a number of other hooks,
the store creates a new listener through Redux's subscribe method. In this
patch we're storing a custom list of listeners before calling into Redux and
relying on a single Redux listener to call all of the registered functions.

This reduces the number of comparisons to `getState()` when updtaing the store
and measured a reduction in several key metrics by between 4% to 9%; specifically
the `type`, `focus`, and `inserter` methods.

Co-authored-by: Jarda Snajdr <jsnajdr@gmail.com>