Show warning on critical block removal (#51145)
* Show warning on critical block removal

* Extract prompt display logic into a hook

* Revert formatting change.

* Prompt for removal of all the blocks

* Move prompt state handling out of BlockList and into self

* findCriticalBlocks: don't dismiss children of matching node

* Refactor by embracing `flatMap` semantics

* When `isBlockCritical( blockName )`, don't immediately return: there
  could be other matching block types within that node's inner blocks.

* findCriticalBlocks -> getBlocksToPromptFor

The intention here was to drop the "find" prefix, which can mislead the
render into believing the function is meant to return a single match or
none.

* Drop isBlockCritical()

* Redesign removal modal

This is an attempt to reconcile a desire for clarity in the UI as to
what is about to be deleted (and why that matters) and the i18n
constraints that apply:

* Remove `__( 'Remove %s?' )` string, which -- although manageable --
  places an added burden on translators of more inflected languages.

* Trim the strings in `blockTypePromptMessages`. This allows us to stack
  all matching blocks' custom messages and finish off with a single
  "Removing it is not advised." paragraph.

* Move prompt into site editor.

* Reset removalPromptExists upon listener unmount

* Let action removeBlocks handle prompts and confirmations

* Add private action `privateRemoveBlocks` to hide extended interface

* Fix unit tests

* Try: Dispatch setRemovalPromptStatus in edit-site init

On one hand, `initializeEditor` is where we would typically set this
sort of setting, which is why this is worth trying out.

On the other hand, interfering with the `REMOVE_BLOCKS` action is a big
deal, so we should be absolutely sure that the prompt status in the
reducer is always accurate. The `useEffect` approach has the advantage
of keeping everything in one place: by rendering a single component like
BlockRemovalWarningModal, we guarantee that both the setting is correct
and the editor will render removal prompts.

* Revert "Try: Dispatch setRemovalPromptStatus in edit-site init"

This reverts commit a5cce0a9ed18ab72c5f37a2df5b355eb7cfa53ad.

I'll repeat my reasoning from the parent commit:

"[...] interfering with the `REMOVE_BLOCKS` action is a big
deal, so we should be absolutely sure that the prompt status in the
reducer is always accurate. The `useEffect` approach has the advantage
of keeping everything in one place: by rendering a single component like
BlockRemovalWarningModal, we guarantee that both the setting is correct
and the editor will render removal prompts."

* Make all actions & selectors private. Rename things.

* Addresses pull request feedback.

* See `FIXME` notes.

* Make BlockRemovalWarningModal private

* Cleanup: Remove BlockList changes from branch

* Tweak removal message for Query. Tweak comments.

* Split action into displayRemovalPrompt & clearRemovalPrompt

The action types have also been renamed: DISPLAY_REMOVAL_PROMPT,
CLEAR_REMOVAL_PROMPT.

Simplifies the function signatures and hopefully makes everything a
little bit clearer.

* Rename setRemovalPromptStatus to toggleRemovalPromptSupport

Also: rename the action type, then sprinkle with comments.

* Rename isRemovalPromptDisplayed to getRemovalPromptData

New reducer: removalPromptData
New selector: getRemovalPromptData

* Add missing @return to displayRemovalPrompt

* Tweak modal copy per feedback

* Turns out private selectors are attached to the thunk proxy!

Well that was easy. :)

* Don't export the new reducers

* Fix tests

---------

Co-authored-by: Miguel Fonseca <150562+mcsf@users.noreply.github.com>