test: Enable jest-watch-typeahead for native tests (#51869)
* test: Enable `jest-watch-typeahead` for native tests

Allow filtering test by file or test name while in watch mode. This
matches existing configuration for the web tests.

* test: Refactor Jest configuration to fix file name filter regex

Previously, Jest was executed from within a nested directory. Doing so
caused the regex patterns used for the `jest-watch-typeahead` to break,
as the `../../` would be erroneously included in the file name.