test: Expand mobile editor tests (#52446)
* test: Test multiple text entry editor history

Expand editor history tests to cover entering text over multiple history
entries.

* test: List split and merge

Expand tests to cover splitting lists and merging with paragraphs.

* test: Highlight text with selection

* test: Update inline comments

The to-do note is not possible for formatted strings that result in
inner HTML tags within the text input. This is because the
`typeInRichText` helper does not support parsing inner HTML tags, only
outer tags. This results in inaccurate cursor placement as the tags
occupy indexes within the string, so the target index is off given it is
based upon the visual position, i.e. the position in the rendered
version of the parsed HTML, rather than the literal position in the
string, i.e. taking into account what positions are occupied by HTML
tags.

If we want to test typing after applying formats, we should likely use
e2e tests.

* test: Remove redundant selectRangeInRichText argument

The helper defaults to passing the second parameter (`start`) as the third parameter (`end`).

Co-authored-by: Carlos Garcia <fluiddot@gmail.com>

* test: Fix editor history test compatibility with new UI

The undo and redo buttons relocated out of the editor and into the host
app in c428fd43302c38a8d0c945452c41235e64bd1607.

* test: Wrap undo and redo actions in act to avoid warnings

The undo and redo actions result in state changes for the editor, which
lead to `act` warnings in the tests. This wraps the actions in `act` so
that warnings are suppressed.

These new undo and redo actions were introduced in
c428fd43302c38a8d0c945452c41235e64bd1607.

---------

Co-authored-by: Carlos Garcia <fluiddot@gmail.com>