Refactor and split navigation block e2e tests into separate files (#52647)
* Move frontend interactivity tests to own file

* Move navigation list view tests to own file

* Add cleanup after link control tests

* Remove unnecessary set-up for navigation block submenu tests

The setup before these tests was already happening in the parent section, so it was running all the setup and teardown twice.

* Fixed create a submenu test incorrect link

This test was adding a navigation to a post, then routing to the homepage to see if the navigation was present there. I think this only worked by accident if the menu happened to be on the homepage. Updated it to go directly to the post instead.

* Consolodate a sole nested test to be top level

* Remove nesting on 'Existing blocks' test

* Use the created post instead of homepage to check frontend result

These were failing on trunk.

* Use Navigation block instead of Navigation block for test names for consistency

* Revert to using / instead of post id path