perf: Reduce unnecessary LinkPicker re-renders (#51613)
* perf: Reduce unnecessary LinkPicker re-renders

Combining state with an object in a single `useState` meant a new object
was set to the state regardless of whether new values were present. This
resulted in unexpected `act` warnings caused by unnecessary re-renders.

Refactoring LinkPicker to separate the two state values into separate
`useState` calls resolves this issue by using primitives that can be
compared directly, which results in `useState` not triggering re-renders
when the values are the same.

* refactor: Simplify component callback invocations

The wrapping anonymous function is unnecessary now that the callback are
simply setting a single state value. Also, passing anonymous functions
to child components can lead to unnecessary re-renders.

* build: Remove erroneous reassure argument

This argument does not exist in the older version of reassure used.
Also, the value is already passed as an environment variable.

* test: Add LinkPicker performance test