Escape on Block Toolbar returns focus to Editor Canvas (#55712)
* Move lastFocus into redux store

Stores the element that had last focus from the editor when focus leaves the editor canvas

* ForwardRef for selected block tools

Adding the escape keypress from the navigable block tools will require forwarding the ref from the BlockTools all the way down to the NavigableToolbar.

* Handle escape on toolbar via toolbar event listener

This method requires forwarding a ref from the editor level down to the navigable toolbar so that the escape unselect shortcut can be blocked and the navigable toolbar event listener will still fire.

Blocking the global escape event shouldn't be necessary, but we have a combination of a few things all combining to create a situation where:
- Because the block toolbar uses createPortal to populate the block toolbar fills, we can't rely on the React event bubbling to hit the onKeyDown listener for the block toolbar
- Since we can't use the React tree, we use the DOM tree which _should_ handle the event bubbling correctly from a `createPortal` element.
- This bubbles via the React tree, which hits this `unselect` escape keypress before the block toolbar DOM event listener has access to it.

Also, this is better than attaching it to all of the children in the block toolbar because when new items are attached to the toolbar (such as via the bubblesVirtually slots or the apply/cancel buttons when cropping an image) we can still catch the events. Otherwise, those buttons are added after the mount, and the children don't receive the listener.

* Refactor to object named notation

* Pass selectedBlockToolsRef to fixed toolbar

* Add keyboard navigation test to cover escape and alt+10 keypresses

* Remove onIndexChange from useEffect dependency array

I added it since it was passed in, but it is a callback and doesn't change. It was causing unnecssary rerenders that was messing with the focus position when selecting an item from a dropdown, such as when changing the heading level on a site title block.