[RNMobile] Add OfflineStatus component (#56934)
* feat: Frame useNetInfo hook foundation

This code is non-functioning currently.

* feat: Add iOS connection status bridge utilities

This bridge will be required for the planned JavaScript Hook to monitor
connection status.

* feat: Add `useIsConnected` hook

Provides React Hook for monitoring the network connection status via the
bridge to the host app.

* Revert "feat: Frame useNetInfo hook foundation"

This reverts commit a8d3660845457787f6b368fe0276bfcfdbd213a6.

* refactor: Align with project Swift syntax

Semicolon is unnecessary.

Co-authored-by: Tanner Stokes <tanner.stokes@automattic.com>

* feat: Add Android connection status bridge utilities

This bridge enables monitoring the connection status on Android.

* feat: Android network connection status request utility

Allow the Android platform to request the current network connection
status.

* fix: Add missing `requestConnectionStatus` bridge method mock

The Demo editor fails to build without a mocked bridge method.

* Add mobile OfflineStatus component

* Add OfflineStatus component to block-list behind __DEV__ flag

* Replace offline icon and update OfflineStatus text alignment

* Update BEM syntax for OfflineStatus

* Update OfflineStatus component colors

* test: Import sole native stylesheet to fix test module resolution error

The current Jest module resolution configuration will fail when there is
only a native-specific file without a web-specific file, e.g. only a
`style.native.scss` and no sibling `style.scss` next to it. Explicitly
importing the native-specific file avoids the error as Jest does not
attempt to seek the non-suffixed file.

* test: Mock `useIsConnected` native bridge method

The native implementation is not available within the JavaScript testing
environment.

---------

Co-authored-by: David Calhoun <github@davidcalhoun.me>
Co-authored-by: Tanner Stokes <tanner.stokes@automattic.com>