Dependency Extraction Webpack Plugin: Fix test issues (#57332)
Webpack errors were not correctly handled the tests. Webpack does not always
pass  an error to the callback, instead it provides errors in the compilation stats.

Here, we check for those errors correctly and tests fail if webpack completes with errors.

Some configuration and test names are also updated in preparation for #57199.