Keep Lock button it in the toolbar until unmounted (#57229)
If the block lock button has been shown, we don't want to remove it from the toolbar until the toolbar is rendered again without a lock button. Removing the lock button beforehand can cause focus loss issues, such as when unlocking the block from the modal. We need to return focus from whence it came, and to do that, we need to leave the button in the toolbar.